refactor: Member 관련 테스트 클래스 위치 변경 #34
